+++ compatibility-tests/scripts/ci.mjs
@@ -1,0 +1,5 @@
+import {startRegistry} from './utils.mjs'
+
+;(async function runCI() {
+  await startRegistry()
+})()
